Mail service, sales table

+++ MLS_Data/Repositories/ColorRepository/ColorRepository.cs
@@ -5,7 +5,7 @@ using MLS_Data.Repositories.GeneralRepository;
 
 namespace MLS_Data.Repositories.ColorRepository
 {
-    public class ColorRepository : GenericRepository<short, Color_DataModel, ApplicationUser_DataModel>, IColorRepository
+    public class ColorRepository : GenericRepository<int, Color_DataModel, ApplicationUser_DataModel>, IColorRepository
     {
         public ColorRepository(MyLittleShopDbContext context) : base(context)
         {
